requirements.txt
streamlit
evaluate
bert-score
pandas
matplotlib
python-dotenv
PyPDF2
requests
